HOLO-133 <refactor: List/Card 코드 분리 및 리팩토링>
