+++ src/app/honeytips/[id]/_components/CommentForm.tsx
@@ -3,6 +3,7 @@
 import { useRef } from "react";
 // import useAuth from "../../_hooks/useHoneytipsAuth";
 import { useAddCommentMutation } from "@/app/honeytips/[id]/_hooks/useCommentMutaion";
+import useAuth from "@/app/honeytips/_hooks/useHoneytipsAuth";
 import type { Post } from "@/app/honeytips/_types/honeytips.type";
 
 type CommentFormProps = {
@@ -12,14 +13,14 @@ type CommentFormProps = {
 const CommentForm = ({ data }: CommentFormProps) => {
   const postId = data.id;
   const inputRef = useRef<HTMLInputElement>(null);
-  // const isAuthenticated = useAuth();
+  const isAuthenticated = useAuth();
   const addMutation = useAddCommentMutation();
 
   const handleSubmit: React.FormEventHandler<HTMLFormElement> = (e) => {
     e.preventDefault();
+
     if (inputRef.current && inputRef.current.value !== "") {
       const newComment = inputRef.current.value;
-      // postId와 newComment를 함께 전달
       addMutation.mutate({ comment: newComment, postId });
       inputRef.current!.value = "";
     }
@@ -33,13 +34,16 @@ const CommentForm = ({ data }: CommentFormProps) => {
       <input
         className="mr-4 flex-grow rounded px-2 py-1 text-sm"
         type="text"
-        placeholder="댓글을 입력해주세요."
+        placeholder={
+          isAuthenticated ? "댓글을 입력해주세요." : "로그인이 필요합니다."
+        }
         ref={inputRef}
+        disabled={!isAuthenticated}
       />
       <button
         className="rounded bg-blue-500 px-2 py-1 text-white hover:bg-blue-700 disabled:cursor-not-allowed disabled:bg-gray-400"
         type="submit"
-        // disabled={!isAuthenticated}
+        disabled={!isAuthenticated}
       >
         저장
       </button>
