+++ src/app/honeytips/[id]/_components/LikeButton.tsx
@@ -4,25 +4,15 @@ import { useLikeMutation } from "@/app/honeytips/[id]/_hooks/useLikeMutation";
 import { useLikeDataQuery } from "@/app/honeytips/[id]/_hooks/useLikeQuery";
 import type { Like } from "@/app/honeytips/_types/honeytips.type";
 import { getId } from "@/app/honeytips/_utils/auth";
-import { useEffect, useState } from "react";
 import { FaHeart, FaRegHeart } from "react-icons/fa";
 
 const LikeButton = ({ postId }: { postId: Like["post_id"] }) => {
-  const [userId, setUserId] = useState<string | null>(null);
   const { data, isError, isPending } = useLikeDataQuery(postId);
   const likeMutation = useLikeMutation(postId);
 
-  useEffect(() => {
-    const fetchUserId = async () => {
-      const id = await getId();
-      setUserId(id);
-    };
-    fetchUserId();
-  }, []);
+  const handleLikeBtn = async () => {
+    const userId: Like["user_id"] | null = await getId();
 
-  console.log('userId', userId)
-
-  const handleLikeBtn = () => {
     if (!userId) {
       alert("로그인이 필요합니다.");
       return;
