+++ src/app/honeytips/_components/PostList.tsx
@@ -1,22 +1,14 @@
 "use client";
 
+import PostCard from "@/app/honeytips/_components/PostCard";
 import useAuth from "@/app/honeytips/_hooks/useHoneytipsAuth";
 import type { Post } from "@/app/honeytips/_types/honeytips.type";
 import { countComments } from "@/app/honeytips/_utils/comment";
 import { countLikes } from "@/app/honeytips/_utils/like";
 import { fetchPostsData } from "@/app/honeytips/_utils/post";
 import clsx from "clsx";
-import dayjs from "dayjs";
-import "dayjs/locale/ko";
-import relativeTime from "dayjs/plugin/relativeTime";
-import Image from "next/image";
-import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
-import { FaRegCommentAlt, FaRegHeart } from "react-icons/fa";
-
-dayjs.extend(relativeTime);
-dayjs.locale("ko");
 
 const PostList = () => {
   const isAuthenticated = useAuth();
@@ -29,12 +21,14 @@ const PostList = () => {
   const [commentsCount, setCommentsCount] = useState<{
     [postId: Post["id"]]: number;
   }>({});
+  const [isLoading, setIsLoading] = useState<boolean>(true); // 로딩 상태 추가
 
   const router = useRouter();
 
   // 마운트 시 전체 포스트 불러오기
   useEffect(() => {
     const fetchPosts = async () => {
+      setIsLoading(true); // 로딩 시작
       const data = await fetchPostsData();
       setPosts(data);
       setFilteredPosts(data);
@@ -64,6 +58,7 @@ const PostList = () => {
 
       setCommentsCount(commentCountMap);
       setLikesCount(likeCountMap);
+      setIsLoading(false);
     };
 
     fetchPosts();
@@ -80,16 +75,6 @@ const PostList = () => {
       setFilteredPosts(filtered);
     }
   }, [selectedCategory, posts]);
-
-  const formatDate = (date: string) => {
-    const now = dayjs();
-    const createdAt = dayjs(date);
-
-    if (now.diff(createdAt, "hour") < 24) {
-      return createdAt.fromNow();
-    }
-    return createdAt.format("YY.MM.DD");
-  };
 
   // 글 작성하기 버튼 클릭 시 로그인 상태 확인
   const handleGoToPost = () => {
@@ -127,67 +112,15 @@ const PostList = () => {
         </button>
       </div>
       <div className="grid grid-cols-1 gap-4">
-        {filteredPosts.length > 0 ? (
+        {isLoading ? (
+          <p className="col-span-full text-center text-gray-500">로딩중...</p>
+        ) : filteredPosts.length > 0 ? (
           filteredPosts.map((post) => (
-            <div key={post.id}>
-              <div className="mb-1 flex items-center justify-between px-1">
-                <div className="flex items-center gap-3">
-                  <Image
-                    className="h-10 w-10 rounded-full border-2"
-                    src={
-                      post.users?.profile_image_url ||
-                      "https://via.placeholder.com/100x100"
-                    }
-                    alt="프로필 이미지"
-                    width={100}
-                    height={100}
-                  />
-                  <p className="font-medium">{post.users?.nickname}</p>
-                </div>
-                <p className="text-xs text-gray-500">
-                  {formatDate(post.created_at)}
-                </p>
-              </div>
-
-              <Link href={`/honeytips/${post.id}`}>
-                <li
-                  className={clsx(
-                    "relative mb-4 flex items-center justify-between gap-2 rounded-lg bg-white p-6 shadow-lg transition-all",
-                    "hover:scale-105 hover:shadow-2xl",
-                  )}
-                >
-                  <div className="flex w-full flex-col">
-                    <h3 className="mb-2 font-bold text-gray-800">
-                      {post.title}
-                    </h3>
-                    <p className="line-clamp-2 text-sm text-gray-600">
-                      {post.content}
-                    </p>
-                    <div className="mt-4 flex items-center gap-1">
-                      <FaRegHeart />
-                      <span className="mr-2 text-xs">
-                        {likesCount[post.id] || 0}
-                      </span>
-                      <FaRegCommentAlt />
-                      <span className="text-xs">
-                        {commentsCount[post.id] || 0}
-                      </span>
-                    </div>
-                  </div>
-                  <Image
-                    className="aspect-square rounded bg-gray-500 object-cover"
-                    src={
-                      post.post_image_url?.[0] ||
-                      "https://via.placeholder.com/120x120"
-                    }
-                    alt="게시글 이미지"
-                    width={120}
-                    height={120}
-                    loading="lazy"
-                  />
-                </li>
-              </Link>
-            </div>
+            <PostCard
+              post={post}
+              likesCount={likesCount}
+              commentsCount={commentsCount}
+            />
           ))
         ) : (
           <p className="col-span-full text-center text-gray-500">
